Aboarding, maybe, excuse me.
Hi, I have an special interest in taking a team-leader/developer role in this project, I think you are in the right way, but with the right support you can make this project good, you have the right ideas but I want to give you a full project scope that we can work on together and make the bots the way you want them to be.

Think about this, I want this project to success as much as you. If you are interested in having me as manager for the project, you can write me back on beno3y@gmail.com, let's chat about this.